pprr/requirements.txt
astroid==2.5.1
colorama==0.4.4
isort==5.7.0
lazy-object-proxy==1.5.2
lxml==4.6.2
mccabe==0.6.1
pylint==2.7.2
python-docx==0.8.10
toml==0.10.2
wrapt==1.12.1
